Separate out time-consuming tests.
