+++ Code/GraphMol/SynthonSpaceSearch/fingerprint_search_catch_tests.cpp
@@ -68,56 +68,6 @@ std::set<std::string> bruteForceSearch(
   return names;
 }
 
-int num5567MReads = 0;
-int num5567RReads = 0;
-
-// The Synthons_5567.csv is in the test data.  It is used multiple times
-// so convert to the binary if that isn't already there.  This makes the
-// tests faster and also tests the conversion process.
-void convert5567FileR() {
-  num5567RReads++;
-  std::string fName(rdbase);
-  std::string libName =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567.csv";
-  std::string binName1 =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567_rdkit.spc";
-  if (!std::filesystem::exists(binName1)) {
-    bool cancelled = false;
-    std::unique_ptr<FingerprintGenerator<std::uint64_t>> fpGen(
-        RDKitFP::getRDKitFPGenerator<std::uint64_t>());
-    convertTextToDBFile(libName, binName1, cancelled, fpGen.get());
-  }
-}
-
-void convert5567FileM() {
-  num5567MReads++;
-  std::string fName(rdbase);
-  std::string libName =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567.csv";
-  std::string binName2 =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567_morgan.spc";
-  if (!std::filesystem::exists(binName2)) {
-    bool cancelled = false;
-    std::unique_ptr<FingerprintGenerator<std::uint64_t>> fpGen(
-        MorganFingerprint::getMorganGenerator<std::uint64_t>(2));
-    convertTextToDBFile(libName, binName2, cancelled, fpGen.get());
-  }
-}
-
-void tidy5567Binary() {
-  std::string fName(rdbase);
-  std::string binName1 =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567_rdkit.spc";
-  std::string binName2 =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567_morgan.spc";
-  if (num5567RReads == 1 && std::filesystem::exists(binName1)) {
-    std::remove(binName1.c_str());
-  }
-  if (num5567MReads == 3 && std::filesystem::exists(binName2)) {
-    std::remove(binName2.c_str());
-  }
-}
-
 TEST_CASE("FP Small tests") {
   REQUIRE(rdbase);
   std::string fName(rdbase);
@@ -182,144 +132,6 @@ TEST_CASE("FP Small tests") {
   }
 }
 
-TEST_CASE("FP Biggy") {
-  REQUIRE(rdbase);
-  std::string fName(rdbase);
-  convert5567FileM();
-  std::string binName =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567_morgan.spc";
-  SynthonSpace synthonspace;
-  synthonspace.readDBFile(binName);
-
-  std::unique_ptr<FingerprintGenerator<std::uint64_t>> fpGen(
-      MorganFingerprint::getMorganGenerator<std::uint64_t>(2));
-
-  std::map<std::string, std::unique_ptr<RWMol>> mols;
-
-  const std::vector<std::string> smis{
-      "c1ccccc1C(=O)N1CCCC1", "c1ccccc1NC(=O)C1CCN1",
-      "c12ccccc1c(N)nc(N)n2", "c12ccc(C)cc1[nH]nc2C(=O)NCc1cncs1",
-      "c1n[nH]cn1",           "C(=O)NC(CC)C(=O)N(CC)C"};
-  const std::vector<size_t> numRes{46, 2, 0, 123, 0, 0};
-  const std::vector<size_t> maxRes{2408, 197, 0, 833, 0, 4};
-  SynthonSpaceSearchParams params;
-  params.approxSimilarityAdjuster = 0.2;
-  params.maxHits = -1;
-  for (size_t i = 0; i < smis.size(); ++i) {
-    auto queryMol = v2::SmilesParse::MolFromSmiles(smis[i]);
-    auto results = synthonspace.fingerprintSearch(*queryMol, *fpGen, params);
-    CHECK(results.getHitMolecules().size() == numRes[i]);
-    CHECK(results.getMaxNumResults() == maxRes[i]);
-  }
-  tidy5567Binary();
-}
-
-TEST_CASE("FP Random Hits") {
-  REQUIRE(rdbase);
-  convert5567FileM();
-  std::string fName(rdbase);
-  std::string binName =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567_morgan.spc";
-  SynthonSpace synthonspace;
-  synthonspace.readDBFile(binName);
-
-  auto queryMol = "c12ccc(C)cc1[nH]nc2C(=O)NCc1cncs1"_smiles;
-  SynthonSpaceSearchParams params;
-  params.maxHits = 100;
-  params.randomSample = true;
-  params.randomSeed = 1;
-  std::unique_ptr<FingerprintGenerator<std::uint64_t>> fpGen(
-      MorganFingerprint::getMorganGenerator<std::uint64_t>(2));
-  auto results = synthonspace.fingerprintSearch(*queryMol, *fpGen, params);
-  std::map<std::string, int> libCounts;
-  for (const auto &m : results.getHitMolecules()) {
-    std::string molName = m->getProp<std::string>(common_properties::_Name);
-    std::string lib(molName.substr(molName.length() - 2));
-    if (const auto &c = libCounts.find(lib); c == libCounts.end()) {
-      libCounts.insert(std::make_pair(lib, 1));
-    } else {
-      c->second++;
-    }
-  }
-  CHECK(results.getHitMolecules().size() == 100);
-  std::map<std::string, int> expCounts{{"a1", 100}};
-  CHECK(expCounts == libCounts);
-  CHECK(results.getHitMolecules().front()->getProp<double>("Similarity") ==
-        Catch::Approx(0.711538));
-  CHECK(results.getHitMolecules().back()->getProp<double>("Similarity") ==
-        Catch::Approx(0.5));
-  tidy5567Binary();
-}
-
-TEST_CASE("Timeout") {
-  REQUIRE(rdbase);
-  convert5567FileM();
-  std::string fName(rdbase);
-  std::string binName =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567_morgan.spc";
-  SynthonSpace synthonspace;
-  synthonspace.readDBFile(binName);
-  SynthonSpaceSearchParams params;
-  params.maxHits = -1;
-  params.similarityCutoff = 0.3;
-  params.fragSimilarityAdjuster = 0.3;
-  params.timeOut = 2;
-  std::unique_ptr<FingerprintGenerator<std::uint64_t>> fpGen(
-      MorganFingerprint::getMorganGenerator<std::uint64_t>(2));
-
-  auto queryMol = "c12ccc(C)cc1[nH]nc2C(=O)NCc1cncs1"_smiles;
-  auto results = synthonspace.fingerprintSearch(*queryMol, *fpGen, params);
-  CHECK(results.getTimedOut());
-
-  // Make sure no timeout also works, but only on a short search.
-  params.maxHits = 100;
-  params.similarityCutoff = 0.3;
-  params.fragSimilarityAdjuster = 0.2;
-  params.timeOut = 0;
-  auto results1 = synthonspace.fingerprintSearch(*queryMol, *fpGen, params);
-  CHECK(!results1.getTimedOut());
-  tidy5567Binary();
-}
-
-TEST_CASE("FP Approx Similarity") {
-  REQUIRE(rdbase);
-  convert5567FileR();
-  std::string fName(rdbase);
-  std::string binName =
-      fName + "/Code/GraphMol/SynthonSpaceSearch/data/Syntons_5567_rdkit.spc";
-  SynthonSpace synthonspace;
-  synthonspace.readDBFile(binName);
-  SynthonSpaceSearchParams params;
-  // The addFP and subtractFP are built from a random selection of
-  // products so do occasionally vary, so use a fixed seed.
-  params.randomSeed = 1;
-  params.similarityCutoff = 0.5;
-  params.timeOut = 0;
-  params.maxHits = 1000;
-
-  std::unique_ptr<FingerprintGenerator<std::uint64_t>> fpGen(
-      RDKitFP::getRDKitFPGenerator<std::uint64_t>());
-  auto queryMol = "c12ccc(C)cc1[nH]nc2C(=O)NCc1cncs1"_smiles;
-
-  // With RDKit fingerprints, 0.05 gives a reasonable compromise
-  // between speed and hits missed.
-  params.approxSimilarityAdjuster = 0.05;
-  auto results = synthonspace.fingerprintSearch(*queryMol, *fpGen, params);
-  CHECK(results.getHitMolecules().size() == 546);
-  CHECK(results.getMaxNumResults() == 1467);
-
-  // A tighter adjuster misses more hits.
-  params.approxSimilarityAdjuster = 0.01;
-  results = synthonspace.fingerprintSearch(*queryMol, *fpGen, params);
-  CHECK(results.getHitMolecules().size() == 187);
-
-  // This is the actual number of hits achievable.
-  params.approxSimilarityAdjuster = 0.25;
-  results = synthonspace.fingerprintSearch(*queryMol, *fpGen, params);
-  CHECK(results.getHitMolecules().size() == 981);
-  tidy5567Binary();
-}
-
 TEST_CASE("FP Binary File") {
   REQUIRE(rdbase);
   std::string fName(rdbase);
@@ -361,63 +173,3 @@ TEST_CASE("Missing exact match") {
   CHECK(results.getHitMolecules().size() == 1);
   CHECK(results.getHitMolecules()[0]->getProp<double>("Similarity") == 1.0);
 }
-
-TEST_CASE("FP Binary File2") {
-  REQUIRE(rdbase);
-  std::string fName(rdbase);
-  SynthonSpace synthonspace;
-  std::string libName =
-      "/Users/david/Projects/SynthonSpaceTests/REAL/2024-09_RID-4-Cozchemix/2024-09_REAL_synthons_rdkit_3000.spc";
-  std::unique_ptr<FingerprintGenerator<std::uint64_t>> fpGen(
-      RDKitFP::getRDKitFPGenerator<std::uint64_t>());
-  synthonspace.readDBFile(libName, -1);
-  CHECK(synthonspace.getNumReactions() == 1008);
-  CHECK(synthonspace.getNumProducts() == 70575407790);
-  std::cout << synthonspace.getNumReactions() << std::endl;
-  std::cout << synthonspace.getNumProducts() << std::endl;
-#if 0
-  SearchResults results;
-  auto queryMol = "O=C(Nc1c(CNC=O)cc[s]1)c1nccnc1"_smiles;
-  CHECK_NOTHROW(results = synthonspace.fingerprintSearch(*queryMol, *fpGen));
-  CHECK(results.getHitMolecules().size() == 211);
-  CHECK(results.getMaxNumResults() == 1397664);
-#endif
-}
-
-#if 0
-// Whilst the code is still under active development, it's convenient to have this
-// in here.  It can come out later.
-TEST_CASE("FP Freedom Space") {
-  std::string libName =
-      "/Users/david/Projects/SynthonSpaceTests/FreedomSpace/2024-09_Freedom_synthons_rdkit.spc";
-  // std::string libName =
-  // "/Users/david/Projects/SynthonSpaceTests/REAL/2024-09_RID-4-Cozchemix/2024-09_REAL_synthons_rdkit_3000.spc";
-  // libName =
-  // "/Users/david/Projects/SynthonSpaceTests/REAL/2024-09_RID-4-Cozchemix/random_real_1_rdkit.spc";
-  SynthonSpace synthonspace;
-  synthonspace.readDBFile(libName);
-  auto m =
-      "C=CC(=O)Nc1cc(Nc2nccc(-c3cn(C)c4ccccc34)n2)c(OC)cc1N(C)CCN(C)C"_smiles;
-  std::unique_ptr<FingerprintGenerator<std::uint64_t>> fpGen(
-      RDKitFP::getRDKitFPGenerator<std::uint64_t>());
-  SynthonSpaceSearchParams params;
-  params.similarityCutoff = 0.4;
-  params.maxHits = 1000;
-  params.fragSimilarityAdjuster = 0.01;
-  params.approxSimilarityAdjuster = 0.05;
-  params.numThreads = -1;
-  SearchResults results;
-  results = synthonspace.fingerprintSearch(*m, *fpGen, params);
-  std::cout << "Number of results : " << results.getHitMolecules().size()
-            << std::endl;
-  int i = 0;
-  for (const auto &mol : results.getHitMolecules()) {
-    if (i < 10 || i > params.maxHits - 10) {
-      std::cout << i << " : " << MolToSmiles(*mol) << " : "
-                << mol->getProp<std::string>(common_properties::_Name) << "  "
-                << mol->getProp<double>("Similarity") << std::endl;
-    }
-    ++i;
-  }
-}
-#endif
